+++ bootique-cayenne-jcache/src/main/java/io/bootique/cayenne/jcache/CayenneJCacheModule.java
@@ -9,10 +9,11 @@ import com.google.inject.Singleton;
 import com.google.inject.binder.LinkedBindingBuilder;
 import com.google.inject.multibindings.Multibinder;
 import io.bootique.cayenne.CayenneModule;
+import io.bootique.cayenne.jcache.cache.TempJCacheConfigurationFactory;
+import io.bootique.cayenne.jcache.cache.TempJCacheQueryCache;
 import io.bootique.jcache.JCacheModule;
 import org.apache.cayenne.cache.QueryCache;
 import org.apache.cayenne.jcache.JCacheConstants;
-import org.apache.cayenne.jcache.JCacheQueryCache;
 import org.apache.cayenne.lifecycle.cache.CacheInvalidationModuleBuilder;
 import org.apache.cayenne.lifecycle.cache.InvalidationHandler;
 
@@ -80,9 +81,11 @@ public class CayenneJCacheModule implements Module {
         return b -> {
             b.bind(CacheManager.class).toInstance(cacheManager);
 
-            // TODO: remove this once Cayenne M6 fixes ordering of JCacheModule auto-loading... for now we must ensure
-            // that JCacheQueryCache is the cache provider
-            b.bind(QueryCache.class).to(JCacheQueryCache.class);
+            // TODO: remove this once Cayenne M6 fixes a few issues:
+            // 1. ordering of JCacheModule auto-loading
+            // 2. support for untyped caches per CAY-2259
+            b.bind(QueryCache.class).to(TempJCacheQueryCache.class);
+            b.bind(TempJCacheConfigurationFactory.class).to(TempJCacheConfigurationFactory.class);
         };
     }