+++ bootique-cayenne-jcache/src/test/java/io/bootique/cayenne/jcache/CayenneJCacheModuleIT.java
@@ -1,5 +1,6 @@
 package io.bootique.cayenne.jcache;
 
+import io.bootique.cayenne.jcache.cache.TempJCacheQueryCache;
 import io.bootique.cayenne.jcache.persistent.Table1;
 import io.bootique.cayenne.test.CayenneTestDataManager;
 import io.bootique.test.BQTestRuntime;
@@ -7,7 +8,6 @@ import io.bootique.test.junit.BQTestFactory;
 import org.apache.cayenne.ObjectContext;
 import org.apache.cayenne.cache.QueryCache;
 import org.apache.cayenne.configuration.server.ServerRuntime;
-import org.apache.cayenne.jcache.JCacheQueryCache;
 import org.apache.cayenne.query.ObjectSelect;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
@@ -16,9 +16,7 @@ import org.junit.Test;
 
 import javax.cache.CacheManager;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertSame;
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.*;
 
 public class CayenneJCacheModuleIT {
 
@@ -42,7 +40,8 @@ public class CayenneJCacheModuleIT {
     @Test
     public void testCacheProvider() {
         QueryCache cache = RUNTIME.getInjector().getInstance(QueryCache.class);
-        assertTrue("Unexpected cache type: " + cache.getClass().getName(), cache instanceof JCacheQueryCache);
+        // TODO: this should be replaced with JCacheQueryCache once TempJCacheQueryCache is removed
+        assertTrue("Unexpected cache type: " + cache.getClass().getName(), cache instanceof TempJCacheQueryCache);
     }
 
     @Test